Major rewrite to load from a file
Instead of being hardcoded and requiring updates to the HTML directly when
changes are desired, it now takes a CSV URL or Google Spreadsheet URL and will
parse that. Besides making the countdowns much easier to update, this also
allows for multiple pages or even users by simply passing in a new URL to the
query param.

The README explains the required format for the CSV/Spreadsheet. There is also
an example.